+++ crates/maybenot-machines/src/tamaraw.rs
@@ -45,6 +45,14 @@ fn make_padding_machine(p: f64) -> Machine {
             start: 0.0,
             max: 0.0,
         },
+        n: Dist {
+            dist: DistType::Uniform {
+                low: 0.0,
+                high: 0.0,
+            },
+            start: f64::MAX,
+            max: 0.0,
+        },
         duration: Dist {
             dist: DistType::Uniform {
                 low: MAX_SAMPLED_DELAY_DURATION,
@@ -215,6 +223,14 @@ fn make_soft_stop_machine(stop_window: f64) -> Machine {
                 high: 0.0,
             },
             start: 0.0,
+            max: 0.0,
+        },
+        n: Dist {
+            dist: DistType::Uniform {
+                low: 0.0,
+                high: 0.0,
+            },
+            start: f64::MAX,
             max: 0.0,
         },
         duration: Dist {